Revert "Clee iteration four"
