+++ lib/customer_repository.rb
@@ -1,16 +1,19 @@
 require_relative 'customer'
-require_relative 'factory'
+require_relative 'data_analyst'
 # This is the CustomerRepository class
 class CustomerRepository
   attr_reader :customers, :parent
-  def initialize(customers, parent)
-    @customers = customers
-    @parent    = parent
-  end
-
-  def self.from_csv(csv, parent)
-    customers = Factory.create_customers(csv, self)
-    CustomerRepository.new(customers, parent)
+  def initialize(csv, parent)
+    @customers = []
+    @parent = parent
+    DataAnalyst.find_customers(csv).each do |attribute|
+      @customers << Customer.new(id:      attribute[0].to_i,
+                                 first:   attribute[1],
+                                 last:    attribute[2],
+                                 created: Time.parse(attribute[3]),
+                                 updated: Time.parse(attribute[4]),
+                                 parent:  self)
+    end
   end
 
   def inspect
