+++ lib/invoice_item_repository.rb
@@ -1,14 +1,14 @@
 require 'bigdecimal'
 require 'time'
 require_relative 'invoice_item'
-require_relative 'factory'
+require_relative 'data_analyst'
 # This is the InvoiceItemRepository class
 class InvoiceItemRepository
   attr_reader :invoice_items, :parent
   def initialize(csv, parent)
     @invoice_items = []
     @parent        = parent
-    Factory.find_attributes(csv).each do |attribute|
+    DataAnalyst.find_invoice_items(csv).each do |attribute|
       @invoice_items << InvoiceItem.new(id:          attribute[0].to_i,
                                         item_id:     attribute[1].to_i,
                                         invoice_id:  attribute[2].to_i,
