+++ lib/invoice_repository.rb
@@ -1,14 +1,14 @@
 require 'bigdecimal'
 require 'time'
 require_relative 'invoice'
-require_relative 'factory'
+require_relative 'data_analyst'
 # This is the InvoiceRepository class
 class InvoiceRepository
   attr_reader :invoices, :parent
   def initialize(csv, parent)
     @invoices = []
     @parent   = parent
-    Factory.find_attributes(csv).each do |attribute|
+    DataAnalyst.find_invoices(csv).each do |attribute|
       @invoices << Invoice.new(id:          attribute[0].to_i,
                                customer_id: attribute[1].to_i,
                                merchant_id: attribute[2].to_i,
