+++ lib/sales_engine.rb
@@ -18,6 +18,6 @@ class SalesEngine
     @invoices      = InvoiceRepository.new(hash[:invoices], self)
     @invoice_items = InvoiceItemRepository.new(hash[:invoice_items], self)
     @transactions  = TransactionRepository.new(hash[:transactions], self)
-    @customers     = CustomerRepository.from_csv(hash[:customers], self)
+    @customers     = CustomerRepository.new(hash[:customers], self)
   end
 end
